separate ConnectionManager.h from ConnectionManager.cc
